Refactoring board component /Add Delivery component / Change display for delivery
