+++ fast-delivery-app/src/component/DeliverymanBoard.js
@@ -1,43 +1,27 @@
 import { useContext, useState, useEffect } from 'react'
 import {
-  Popover,
-  PopoverTrigger,
-  PopoverBody,
-  PopoverArrow,
-  PopoverCloseButton,
-  PopoverContent,
   Text,
   HStack,
   Box,
   Flex,
   Button,
   Heading,
-  Input,
-  useToast,
   Spacer,
-  Divider,
-  useDisclosure
 } from '@chakra-ui/react'
-import {
-  Modal,
-  ModalContent,
-  ModalHeader,
-  ModalFooter,
-  ModalBody,
-  ModalCloseButton,
-} from "@chakra-ui/react"
+
+import DeliverymanDelivery from './DeliverymanDelivery'
 import { Web3Context } from 'web3-hooks'
 import { FastDeliveryNftContext } from '../App'
 import { FastDeliveryUserContext } from '../App'
-import { ethers, utils } from 'ethers'
+import { ethers } from 'ethers'
 
 function DeliverymanBoard(props) {
   const { selectedId, setSelectedId } = props
+
   const [web3State] = useContext(Web3Context)
   const fastDeliveryNft = useContext(FastDeliveryNftContext)
   const fastDeliveryUser = useContext(FastDeliveryUserContext)
-  const toast = useToast()
-  const { isOpen, onOpen, onClose } = useDisclosure()
+
 
 
   const [lastDeliveryId, setLastDeliveryId] = useState()
@@ -45,11 +29,9 @@ function DeliverymanBoard(props) {
   const [loadingList, setLoadingList] = useState(false)
   // eslint-disable-next-line no-unused-vars
   const [isLoading, setIsLoading] = useState(false)
-  const [idSelect2, setIdSelect2] = useState("")
-  const [deliveryCode, setDeliveryCode] = useState("")
   const [displayAddDeliveryForDeliveryMan, setDisplayAddDeliveryForDeliveryMan] = useState(false)
   const [selectFilter, setSelectFilter] = useState(ethers.utils.getAddress(web3State.account))
-  const [isChanging, setIsChanging] = useState(false)
+  const [deliveryCode, setDeliveryCode] = useState("")
 
 
   // get last Nft number id
@@ -77,6 +59,7 @@ function DeliverymanBoard(props) {
       3: "delivered",
       4: "deleted"
     }
+    let list = []
     let listHeader = {
       id: "Id",
       senderAddress: "Sender",
@@ -86,11 +69,11 @@ function DeliverymanBoard(props) {
       deliveryStatus: "Status",
       timestamp: "Date"
     }
-    let list = []
     if (fastDeliveryNft) {
       const getAllDeliveries = async () => {
-        setLoadingList(true)
         try {
+          list = []
+          setLoadingList(true)
           let deliveryDateStatus
           let deliveryStatusInfo
           for (let i = 1; i <= lastDeliveryId; i++) {
@@ -157,13 +140,14 @@ function DeliverymanBoard(props) {
           console.log(e)
         } finally {
           setLoadingList(false)
-          setIsChanging(false)
         }
       }
       const cb = (Sender, deliveryman, tokenId) => {
+        list = []
         getAllDeliveries()
       }
-      if (!isChanging) { getAllDeliveries() }
+      getAllDeliveries()
+
       const inDeliveryFilter = fastDeliveryNft.filters.InDelivery(null, web3State.account)
       const attributedFilter = fastDeliveryNft.filters.Attributed(null, web3State.account)
       const deliveredFilter = fastDeliveryNft.filters.Delivered(null, web3State.account)
@@ -178,89 +162,11 @@ function DeliverymanBoard(props) {
         fastDeliveryNft.off(deliveredFilter, cb)
       }
     }
-  }, [fastDeliveryNft, fastDeliveryUser, lastDeliveryId, setIsChanging, web3State.account, isChanging])
+  }, [fastDeliveryNft, web3State.account, lastDeliveryId, fastDeliveryUser])
 
   const handleClickDisplayDelivery = () => {
     setDisplayAddDeliveryForDeliveryMan(!displayAddDeliveryForDeliveryMan)
-    setIdSelect2("")
-    setDeliveryCode("")
     selectFilter === ethers.constants.AddressZero ? setSelectFilter(ethers.utils.getAddress(web3State.account)) : setSelectFilter(ethers.constants.AddressZero)
-  }
-
-  // add A delivery for deliveryman
-  const handleClickChooseDelivery = async (index) => {
-    try {
-      setIsLoading(true)
-      let tx = await fastDeliveryNft.attributeDelivery(index)
-      await tx.wait()
-      toast({
-        title: `Confirmed transaction : delivery id ${index} have been added to your deliveries`,
-        description: `Transaction hash: ${tx.hash}`,
-        status: 'success',
-        duration: 5000,
-        isClosable: true,
-      })
-    } catch (e) {
-      if (e.code) {
-        console.log(e.message)
-        toast({
-          title: 'Transaction denied',
-          description: e.message,
-          status: 'error',
-          duration: 9000,
-          isClosable: true,
-        })
-      }
-      console.log(e)
-    } finally {
-      setIsLoading(false)
-      setIdSelect2("")
-      setIsChanging(true)
-    }
-  }
-
-  // confirm a parcel is delivered by deliveryman
-  const handleClickConfirmDelivered = async (index) => {
-    console.log(deliveryCode, 'Delivery code')
-    const deliveryInfo = await fastDeliveryNft.DeliveryInfo(index)
-    try {
-      setIsLoading(true)
-      let tx = await fastDeliveryNft.delivered(index, utils.keccak256(utils.toUtf8Bytes(deliveryCode.toString())))
-      await tx.wait()
-      toast({
-        title: `Confirmed transaction : delivering for id ${index} has been registered \n Delivery id ${index} is payed`,
-        description: `Transaction hash: ${tx.hash}`,
-        status: 'success',
-        duration: 5000,
-        isClosable: true,
-      })
-    } catch (e) {
-      if (utils.keccak256(utils.toUtf8Bytes(deliveryCode.toString())) !== deliveryInfo.deliveryCode) {
-        toast({
-          title: 'Transaction denied',
-          description: `Delivery code is false for Id ${idSelect2}`,
-          status: 'error',
-          duration: 9000,
-          isClosable: true,
-        })
-      }
-      if (e.code) {
-        console.log(e.message)
-        toast({
-          title: 'Transaction denied',
-          description: e.message,
-          status: 'error',
-          duration: 9000,
-          isClosable: true,
-        })
-      }
-      console.log(e)
-    } finally {
-      setIsLoading(false)
-      setIdSelect2("")
-      setDeliveryCode("")
-      setIsChanging(true)
-    }
   }
 
   /* no implemented
@@ -297,7 +203,7 @@ function DeliverymanBoard(props) {
   return (
     <>
       <Box w="100%" px="2" py="2">
-        <HStack pl="1" py="2" position="sticky" top="125px" zIndex="sticky" alignItems="center" justifyContent="space-beetween" w="100%">
+        <HStack pl="1" py="1" position="sticky" top="125px" zIndex="sticky" alignItems="center" justifyContent="space-beetween" w="100%">
           <Heading size="md">{!displayAddDeliveryForDeliveryMan === true ? 'Deliveries' : 'Add Delivery'}</Heading>
           <Spacer />
           <Flex wrap="wrap" justifyContent="flex-end" alignItems="center">
@@ -326,148 +232,8 @@ function DeliverymanBoard(props) {
             )
               .map(delivery => {
                 return (
-                  <Box p="0" w="100 % " key={delivery.id} border={isNaN(delivery.id) ? '1px' : ''}>
-                    < Flex wrap="wrap" mt={isNaN(delivery.id) ? '0' : '2'}>
-                      <Box display="flex" alignItems="center" justifyContent="center" bg="gray.200" w="25px" px="1">{delivery.id}</Box>
-                      <Box display="flex" mx="1" bg="blue.300" minW="250px">
-                        <Popover isLazy trigger="hover" >
-                          <PopoverTrigger>
-                            <Flex mx="1" wrap="wrap" alignItems="center" justifyContent="center" maxW="240px">
-                              {delivery.senderAddress}</Flex>
-                          </PopoverTrigger>
-                          <PopoverContent mx="1">
-                            <PopoverArrow />
-                            <PopoverCloseButton />
-                            <PopoverBody>{delivery.recipientFirstName} {delivery.recipientLastName}</PopoverBody>
-                            <PopoverBody>{delivery.senderAddress}</PopoverBody>
-                            <PopoverBody>{delivery.senderInfo}</PopoverBody>
-                            <PopoverBody>{delivery.senderTel}</PopoverBody>
-                            <PopoverBody>{delivery.senderMail}</PopoverBody>
-                          </PopoverContent>
-                        </Popover>
-                      </Box>
-                      <Box display="flex" mx="1" bg="teal.300" minW="250px">
-                        <Popover isLazy trigger="hover" >
-                          <PopoverTrigger>
-                            <Flex mx="1" wrap="wrap" alignItems="center" justifyContent="center" maxW="240px">
-                              {delivery.recipientAddress}</Flex>
-                          </PopoverTrigger>
-                          <PopoverContent mx="1">
-                            <PopoverArrow />
-                            <PopoverCloseButton />
-                            <PopoverBody>{delivery.recipientFirstName} {delivery.recipientLastName}</PopoverBody>
-                            <PopoverBody>{delivery.recipientAddress}</PopoverBody>
-                            <PopoverBody>{delivery.recipientInfo}</PopoverBody>
-                            <PopoverBody>{delivery.recipientTel}</PopoverBody>
-                            <PopoverBody>{delivery.recipientMail}</PopoverBody>
-                          </PopoverContent>
-                        </Popover>
-                      </Box>
-                      <Box display="flex" mx="1" alignItems="center" justifyContent="center" minW="90px">
-                        {!isNaN(delivery.deliveryDistance) ? (Math.round(delivery.deliveryDistance * 100) / 100) : 'D'} (Km)
-                      </Box>
-                      <Box display="flex" mx="1" alignItems="center" justifyContent="center" minW="100px" >
-                        {!isNaN(delivery.deliveryAmount) ? (Math.round(delivery.deliveryAmount * 100) / 100) : 'Price'} (DAID)
-                      </Box>
-                      <Popover isLazy trigger="hover" >
-                        <PopoverTrigger>
-                          <Box display="flex" flex="1" alignItems="center" justifyContent="center" bg="orange.200" minW="150px">
-                            {delivery.deliveryStatus}
-                          </Box>
-                        </PopoverTrigger>
-                        <PopoverContent mx="1">
-                          <PopoverArrow />
-                          <PopoverCloseButton />
-                          <PopoverBody>{delivery.deliveryStatus}</PopoverBody>
-                          <PopoverBody>{delivery.timestamp}</PopoverBody>
-                        </PopoverContent>
-                      </Popover>
-                      {delivery.deliveryStatus === "onLine" && (
-                        <Box display="flex" mx="1" alignItems="center" justifyContent="center">
-                          <Button
-                            m="1"
-                            p="2"
-                            size="sx"
-                            borderRadius="lg"
-                            colorScheme="blue"
-                            type="button"
-                            onClick={() => {
-                              onOpen()
-                              setSelectedId(Number(delivery.id))
-                            }
-                            }>Add</Button>
-                          <Modal isOpen={isOpen} onClose={onClose}>
-                            <ModalContent>
-                              <ModalHeader>Add delivery confirmation</ModalHeader>
-                              <ModalCloseButton />
-                              <ModalBody >
-                                <Text>Do you want to add delivery : ID {(selectedId)}?</Text>
-                              </ModalBody>
-                              <ModalFooter>
-                                <Button colorScheme="blue" mr={3} onClick={onClose}>
-                                  Close
-                                </Button>
-                                <Button
-                                  colorScheme="blue"
-                                  mr={3}
-                                  onClick={() => {
-                                    onClose()
-                                    handleClickChooseDelivery(selectedId)
-                                  }}
-                                  variant="ghost">Ok</Button>
-                              </ModalFooter>
-                            </ModalContent>
-                          </Modal>
-                        </Box>
-                      )}
-                      {delivery.deliveryStatus === "inDelivery" && (
-                        <Box display="flex" mx="1" alignItems="center" justifyContent="center">
-                          <Button
-                            m="1"
-                            p="2"
-                            size="sx"
-                            borderRadius="lg"
-                            colorScheme="orange"
-                            type="button"
-                            onClick={() => {
-                              onOpen()
-                              setSelectedId(Number(delivery.id))
-                            }
-                            }>ID {Number(delivery.id)} delivered</Button>
-                          <Modal isOpen={isOpen} onClose={onClose}>
-                            <ModalContent>
-                              <ModalHeader>Delivery confirmation</ModalHeader>
-                              <ModalCloseButton />
-                              <ModalBody >
-                                <Text>Please enter delivery code for delivery ID: {(selectedId)}</Text>
-                                <Input
-                                  w="85px"
-                                  pl="1"
-                                  type="number"
-                                  placeholder="Code"
-                                  onChange={(event) => setDeliveryCode(event.target.value)}
-                                />
-                              </ModalBody>
-                              <ModalFooter>
-                                <Button colorScheme="blue" mr={3} onClick={onClose}>
-                                  Close
-                                </Button>
-                                <Button
-                                  colorScheme="blue"
-                                  mr={3}
-                                  onClick={() => {
-                                    onClose()
-                                    handleClickConfirmDelivered(selectedId)
-                                  }}
-                                  variant="ghost">Ok</Button>
-                              </ModalFooter>
-                            </ModalContent>
-                          </Modal>
-                        </Box>
-                      )
-                      }
-                    </Flex>
-                    {!isNaN(delivery.id) && (<Divider />)}
+                  <Box p="0" mt="2" w="100%" key={delivery.id} border="1px">
+                    <DeliverymanDelivery delivery={delivery} deliveryCode={deliveryCode} setDeliveryCode={setDeliveryCode} selectedId={selectedId} setSelectedId={setSelectedId} />
                   </Box>
                 )
               })
